Develop

+++ Core/Domain/src/main/java/ir/otragh/core/domain/homes/entities/Home.java
@@ -1,5 +1,6 @@
 package ir.otragh.core.domain.homes.entities;
 
+import ir.otragh.core.domain.framework.Result;
 import ir.otragh.core.domain.framework.entities.BaseEntity;
 import ir.otragh.core.domain.homes.events.HomeCreated;
 import ir.otragh.core.domain.homes.events.HomeReserved;
@@ -13,7 +14,7 @@ import java.time.LocalDateTime;
 import java.util.LinkedList;
 
 @Getter
-public final class Home extends BaseEntity<Integer> {
+public final class Home extends BaseEntity<Long> {
     private Title title;
     private Description description;
     private Address address;
@@ -21,7 +22,7 @@ public final class Home extends BaseEntity<Integer> {
     private LinkedList<Long> amenities;
     private LocalDateTime lastReservedOnUTC;
 
-    private Home(int id, Title title, Description description, Address address, Money price, LinkedList<Long> amenities) {
+    private Home(long id, Title title, Description description, Address address, Money price, LinkedList<Long> amenities) {
         super(id);
         this.title = title;
         this.description = description;
