+++ Core/Domain/src/main/java/ir/otragh/core/domain/homes/events/HomeReserved.java
@@ -2,5 +2,5 @@ package ir.otragh.core.domain.homes.events;
 
 import ir.otragh.core.domain.framework.events.DomainEvent;
 
-public record HomeReserved(int id) implements DomainEvent {
+public record HomeReserved(long id) implements DomainEvent {
 }
